+++ src/specific/s_file.h
@@ -8,6 +8,5 @@
 bool LoadLevel(const char *filename, int32_t level_num);
 bool S_LoadLevel(int32_t level_num);
 const char *GetFullPath(const char *filename);
-void FileLoad(const char *path, char **output_data, size_t *output_size);
 
 #endif
